% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stuff.R
\name{add_stuff}
\alias{add_stuff}
\title{Add stuff}
\usage{
add_stuff(first, second)
}
\arguments{
\item{first}{First this to add.}

\item{second}{Second thing to add.}
}
\value{
first plus second thing
}
\description{
Add stuff
}
\details{
Add stuff
}

