% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_puller.R
\name{ds_puller}
\alias{ds_puller}
\title{ds_puller}
\usage{
ds_puller(SECURITITES, FIELDS, fromDATE, toDATE, PERIOD, CURRENCY = NULL,
  REQUESTS_OFF = F)
}
\arguments{
\item{SECURITITES}{A Datastream code.}

\item{FIELDS}{A Datastream field.}

\item{fromDATE}{Data start date.}

\item{toDATE}{Data end date.}

\item{PERIOD}{Data Frequency: "D", "M", "Q," "Y".}

\item{CURRENCY}{"USD", "EUR", etc. or NULL for default.}

\item{REQUESTS_OFF}{Should requests not be used?}
}
\value{
a list with two dataframes, named Data and Metadata
}
\description{
A function for pulling Datastream data.
}
\details{
This function is a wrapper for the function ds() from
from the package Rdatastream.
}
\examples{
ds_puller( SECURITITES = c("TOTMKUS"), FIELDS = "MV", fromDATE = "2014-09-14", toDATE = "2015-01-31", PERIOD = "D", CURRENCY = "EUR" )

}

