% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_puller.R
\name{ds_puller}
\alias{ds_puller}
\title{ds_puller}
\usage{
ds_puller(SECURITITES, FIELDS, fromDATE, toDATE, PERIOD, CURRENCY = NULL)
}
\arguments{
\item{SECURITITES}{A vector of Datastream codes.}

\item{FIELDS}{A vector of Datastream fields.}

\item{fromDATE}{Date from which to begin data pull.}

\item{toDATE}{Date to end data pull.}

\item{PERIOD}{Frequency of data: "D", "M", "Q," "Y".}

\item{CURRENCY}{Convert to "USD", "EUR", etc? If NULL default currency is used.}
}
\value{
a list with two dataframes, named Data and Metadata
}
\description{
A function for pulling Datastream data
}
\details{
This function is a wrapper for the function ds() from
from the package Rdatastream.
}
\examples{
ds_puller( SECURITITES = c("TOTMKUS"), FIELDS = list("MV"), fromDATE = "2014-09-14", toDATE = "2015-01-31", PERIOD = "D")

}

